requirements.txt
# Travel Agent - Production
google-adk>=0.2.0
google-genai>=0.5.0
fastapi>=0.115.0
uvicorn>=0.32.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-dotenv>=1.0.0
requests>=2.31.0
pytz>=2024.1
firebase-admin
redis>=5.0.0
slowapi>=0.1.9
